notion_sync_read/requirements.txt
notion_client
python-dotenv
python-decouple
python-slugify